+++ content.js
@@ -2,6 +2,7 @@ chrome.storage.sync.get( ["mixinsScript", "mixinsState"], stored =>
 {
 	if( stored.mixinsState == "mixinsDisabledState" ) return;
 	
+	
 	// Use in mixins which run as content script in order to avoid XSS.
 	// ECMAScript provides URI percent-encoding routines only, 
 	// so we have to define our own HTML-entities encoder:
@@ -10,33 +11,58 @@ chrome.storage.sync.get( ["mixinsScript", "mixinsState"], stored =>
 		return theStr.replace( /[\u00A0-\u99999<>\&]/gim, (i) => '&#' + i.charCodeAt( 0 ) + ';' );
 	}
 	
+	
 	function redir( theRegex, theReplace ) { /* implemented in background.js */ };
 	
-	function mixin( theUrl, theCallback, theOpts )
-	{
-		if( !location.href.startsWith( theUrl ) ) return;
+	
+	function mixin( theUrls, theCode, theOpts )
+	{	
+ 		var urls;  // One or multiple URLs given
 		
-		// Needs more privileges, e.g., if CORS issues:
-		if( theOpts && (theOpts.runAsContentScript || false) )  
+		if( typeof theUrls === "string" || theUrls instanceof String )
+			urls = [ theUrls ];
+			
+		if( Array.isArray( theUrls ) )
+			urls = theUrls;
+				
+		if( !urls || !urls.some( u => location.href.startsWith( u ) ) )
+			return;
+		
+		if( typeof theCode === "string" || theCode instanceof String )  // Inject CSS:
 		{
-			theCallback();
+			const s = document.createElement( "style" );
+			s.textContent = theCode;
+			( document.head || document.documentElement ).appendChild( s );
 			return;
 		}
 		
-		// Content scripts and the website share the DOM but no JS functions.
-		// https://developer.chrome.com/extensions/content_scripts#isolated_world
-		// We run user scripts in the world of the target website.
-		// This allows us to call the website's Javascript functions
-		// and to strip off some privileges: the remote page cannot access
-		// our extension.
+		if( typeof theCode === "function" )  // Inject ECMAScript:
+		{
+			// Content scripts and the website share the DOM but no JS functions.
+			// https://developer.chrome.com/extensions/content_scripts#isolated_world
+			// But we can run user scripts in the world of the target website.
+			// This allows us to call the website's Javascript functions and to strip
+			// off some privileges: the remote page cannot access our extension.
+			
+			// Some code, however, needs more privileges, e.g., on CORS issues:
+			if( theOpts && (theOpts.runAsContentScript || false) )
+			{
+				theCode();
+				return;
+			}
+			
+			const s = document.createElement( "script" );
+			s.textContent = "(" + theCode + ")();";
+			( document.head || document.documentElement ).appendChild( s );
+			s.remove();  // ??
+			return;
+		}
 		
-		const s = document.createElement( "script" );
-		s.textContent = "(" + theCallback + ")();";
-		( document.head || document.documentElement ).appendChild( s );
-		s.remove();
+		throw "Given code neither string nor function";
 	};
 	
-	
+
 	eval( stored.mixinsScript );  // Script calls mixin(), redir() multiple times
+	                              // and prints SyntaxError.message to console
 });