+++ CHANGELOG.md
@@ -4,9 +4,19 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org)
 
 
+
+## [0.3.1] - 2018-08-28
+
+New: 
+	- pass multiple URLs to `mixin(...)`
+	- inject CSS code instead of Javascript by passing a template literal to `mixin(...)`
+	  (see script-example.js)
+
+
 ## [0.3.0] - 2018-08-17
 
-New: Redirect requests with `redir( ... )`
+New: 
+	- redirect requests with `redir( ... )`
 
 
 ## [0.2.0] - 2018-05-27